pyproject.toml
[project]
name = "polars-quant"
description = "Blazingly Fast Polars Quant Tool"
readme = "README.md"
authors = [
    { name = "Firstastor", email = "23150126@hdu.edu.cn" }
]
version = "1.0.0"
license = { file = "LICENSE" }
requires-python = ">=3.10"

dependencies = [
    "polars>=1.34.0",
    "pyarrow>=23.0.1",
]

[build-system]
requires = ["maturin>=1.10,<2.0"]
build-backend = "maturin"

[tool.maturin]
python-source = "python"

[tool.uv]
package = false

[dependency-groups]
dev = [
    "baostock>=0.8.9",
    "numpy>=2.2.6",
    "pytest-benchmark>=5.2.3",
    "pytest-memray>=1.8.0",
    "ta-lib>=0.6.8",
]
